FallDetector_Watch/ml/requirements.txt
pyserial>=3.5
tensorflow>=2.12
numpy>=1.23
pandas>=1.5
scikit-learn>=1.2
matplotlib>=3.6
